damagability is now part of WARUnit
