--- Assets/Engine/Units/Scripts/WARDamagable.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: cfebe50439bc540bfb420460dd2ba058
-timeCreated: 1501399616
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
