+++ Assets/Resources/Armies/Shmoogaloo/Troop/ShmooTroop.prefab
@@ -172,6 +172,7 @@ GameObject:
   m_Component:
   - component: {fileID: 4442153901002080}
   - component: {fileID: 114619833002364076}
+  - component: {fileID: 114531634033334682}
   m_Layer: 0
   m_Name: ShmooTroop
   m_TagString: Untagged
@@ -565,6 +566,18 @@ MeshCollider:
   m_InflateMesh: 0
   m_SkinWidth: 0.01
   m_Mesh: {fileID: 4300002, guid: ba009f03bbe6cb54298666b0d3b99f5b, type: 3}
+--- !u!114 &114531634033334682
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 1549523744643522}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: cfebe50439bc540bfb420460dd2ba058, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  maxHealth: 10
 --- !u!114 &114619833002364076
 MonoBehaviour:
   m_ObjectHideFlags: 1
@@ -576,6 +589,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: d4b8647674a7f432d88012c117bb5d39, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  speed: 0.5
+  owner: 0
 --- !u!198 &198053891765983028
 ParticleSystem:
   m_ObjectHideFlags: 1
